requirements.txt
asgiref==3.4.1
Django==3.2.5
django-filter==2.4.0
djangorestframework==3.12.4
Markdown==3.3.4
pytz==2021.1
gunicorn==20.1.0
sqlparse==0.4.1
django-adminlte3==0.1.6
django-cors-headers==3.7.0
djangorestframework-simplejwt==4.7.2